pyproject.toml
[tool.poetry]
name = "product-matcher"
version = "0.1.0"
description = "tool to match products across different eCommerce retailer to enhance the Product Catalog management for eCommerce sellers"
authors = ["suri chen <sc4418@columbia.edu>"]
readme = "README.md"
packages = [{include = "product_matcher"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pandas = "^1.5.3"
scikit-learn = "^1.2.1"
nltk = "^3.8.1"
spacy = "^3.5.0"
jupyter = "^1.0.0"
notebook = "^6.5.2"
sentence-transformers = "^2.2.2"
matplotlib = "^3.6.3"
hydra-core = "^1.3.1"
plotly = "^5.13.0"
tensorflow = "^2.11.0"


[tool.poetry.dependencies.en_core_web_sm]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz"
[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


